minor prompt tweak, start of de-slopping
